Remove one level of nesting from in-originals-pred test
in-originals-pred should always be passed a collection of symbols,
libspecs, and prefix lists:

        (in-originals-pred '[java.util.Date
                             (java.util Random BitSet)
                             [clojure.string :as string]
                             [clojure.java [io :as io]]]

In no case will the first member of a properly formatted libspec or
prefix list be a collection itself, so ['((java.util Date UUID))'] is
invalid input.

This test currently passes regardless because the in-originals-pred
implementation does not do any destructuring.