Reimplement in-originals-pred with expand-libs
Previously, reconstructing the following namespace:

        (ns alpha.core
          (:require [alpha.beta :as beta]
                    [alpha.gamma :as gamma]))

        (beta/hello)
        (gamma/hello)

returned the following erroneous ns form:

        (ns alpha.core
          (:require [alpha.beta :as beta]
                    [alpha.beta :as gamma]))

This was due to the implementation of in-originals-pred, which attempted
to determine if a candidate libspec was present in the original require
by regex matching only the prefix of the candidate libspec against each
original require as a string.

e.g.
        ((in-originals-pred '([alpha.beta :as beta]
                              [alpha.gamma :as gamma]))
           '[alpha.beta :as gamma])

returns a true value even though [alpha.beta :as gamma] is not in the
original forms.

This commit reimplements in-originals-pred so that candidate forms are
matched exactly against the fully expanded set of original requires.