Avoid excessive tempfile creation
Use of .deleteOnExit with tempfiles is convenient and has no practical
downsides when running slamhound from the command line.

However, when used from a persistent REPL the number and size of
tempfiles can grow to a non-negligible quantity. For instance, many
Linux distributions use a ramdisk (tmpfs) for /tmp; running slamhound on
a large file repeatedly can quickly consume megabytes of RAM.