regrow/candidates now return sets of symbols
Instead of returning a form for direct insertion, #'candidates now
returns a set of symbols representing the classes or namespaces that
satisfy the given constraints.

The :require-as and :require-refer keys have been changed to :alias and
:refer respectively to match the keys in the new nsrefs map.

The `missing` parameter is now always a symbol as it is more convenient
when working with the keys from ns-publics.

Finally, #'candidates is now public with its own unit tests.