Add regrow/filter-excludes and regrow/filter-renames
We can now use the old ns data to determine if the user wishes to
exclude or rename some namespaces's vars.

This is particularly useful for libraries like core.logic that recommend
replacing many vars in clojure.core; it is annoying when slamhound
inserts (:require [clojure.core :refer [==]]) when you intended it to be
excluded.