Extract missing-sym-patterns to avoid repetition
Visually diffing the regular expressions in #'missing-sym-name was
difficult because they were not aligned to the same column.

Commit 56b5b6b5624673b859c878ae851161f50bcd9f5d added '\+' to one
pattern for vars, but not to the others. This broke with the history of
missing-sym-name, which has used the same pattern for matching vars and
namespaces since its introduction.

We restore this symmetry in this commit for the sake of uniformity.

While we're touching this pattern, add a test for finding
non-alphanumeric vars.