Find special-case token "/" during regrow
Although / is used as the namespace qualifier, the Clojure reader does
treat a lone / as a valid token for the sake of the division function.

Therefore namespaces can also declare their own vars named /, so we
should attempt to find them.

Note that clojure.core// is also special cased; my.ns// is invalid.