Finish support for multiple :require clauses
We added support for handling multiple require clauses in
asplode/ns-to-map (a2fcfdd5a1d2109308503db46a10c7441d692eea), but the
clauses were still being poured into a single list. The effect of this
was that a single :flag would apply to all namespaces, which is what we
were trying to avoid.

ns-to-map now conjes new require clauses, instead of pouring them with
into. asplode/parse-ns-map now correspondingly reduces a list of a list
of libspecs, instead of just a list of libspecs.