Update README with more suggestions for syntax quoted references
It apparently has always been possible (at least as far back as the
history of the Clojure git repo shows) to attach metadata to defn and
defmacro vars by providing a map _after_ the name:

author Rich Hickey <richhickey@gmail.com> Tue Jul 29 16:46:44 2008 +0000

(def

 #^{:doc "Like defn, but the resulting function name is declared as a
  macro and will be used as a macro by the compiler when it is
  called."
    :arglists '([name doc-string? attr-map? [params*] body]
                [name doc-string? attr-map? ([params*] body)+ attr-map?])}
 defmacro (fn [name & args]
            (list 'do
                  (cons `defn (cons name args))
                  (list '. (list 'var name) '(setMacro)))))

This form of setting metadata is cleaner and easier to remember than the
^{} version, so we amend the example in the README.

Also, if the quoted vars refer to functions or constant data, unquoting
them also forces compile-time resolution, so the metadata trick is not
necessary at all in those cases.

While we're here, we also change the screencast URL to the canonical
version from the vanity URL that Vimeo seems to support half-heartedly.