Memoize expensive ns data transformations in a dynamic var
Namespaces often contain more than one import/alias/refer. In these
cases the repeated data transformations of (all-ns) in a single ns
reconstruction is wasteful. Properly memoizing these calculations
results in a considerable performance gain for namespaces with many
references.

A dynamic var is a suitable place for a cache:

    - The current compiler-failure algorithm forces us to work
      sequentially, so a thread binding is not a problem

    - Reconstruction _should_ work with and without a cache for testing

A parallel reconstruction algorithm will require explicit
parameterization of the cache, but we can cross that bridge when we get
there.

The *cache* var is not bound in slam.hound.regrow; that is up to the
discretion of consuming code (tests and the main entry points in
slam.hound). The with-regrow-cache macro can be nested, in which case
the topmost cache is used.

The performance benefit of this change depends on the namespace:

    - Namespaces with a single import/alias/refer are unaffected¹

    - Small namespaces with multiple references of each type experience
      relatively large gains (40% - 50% reduction of reconstruction time)

    - Large namespaces with multiple references experience the same
      absolute speedup as above, but since the time required for
      evaluation of the namespace now dominates, the relative reduction
      in reconstruction time is less dramatic (10% - 30%)

It is unclear whether the size and use of this cache has negative
effects on GC, so this branch is considered experimental. Intuitively,
the reduction of object creation should help reduce GC hiccups, but
tests should be done in larger projects.

The overall effect of this change is that ns evaluation in
check-for-failure is now the largest performance bottleneck by two
orders of magnitude.

¹ Actually, commit d80d4a3 removed the inner loop in :refer candidate
  generation, but the new symbols->ns-syms re-adds one. The inner loops
  of both iterate through the public vars of each ns, so running time is
  only increased linearly by the average count of public vars per ns *
  ns, not geometrically by the number of namespaces.