Execute reconstruct and -main with-regrow-cache
The two main entry points for Slamhound are slam.hound/reconstruct and
slam.hound/-main. Binding slam.hound.regrow/*cache* in both places
allows speed up of both single-file reconstruction (via an editor),
as well as an even greater speed-up of recursive reconstruction of a
directory.

One issue that may arise is that all data about namespaces is cached
on first calculation, so any namespaces or classes created as a result
of a newly evaluated namespace will not be visible in subsequent
reconstructions while reconstructing all clojure files in a directory.

It's unclear how utilizing this assumption about the behavior of
recursive reconstruction could be regarded as anything but a bug, so I
am unconcerned.