Handle case where an alias is used to refer to private vars
e.g.
    (ns alpha)

    (def ^:private foo)
    (def bar)

    (ns beta)

    (def foo #'a/foo)
    (def bar a/bar)

The beta namespace uses the `a` alias to refer to both alpha/bar and the
private var alpha/foo.

Slamhound previously failed to find `alpha` -> `a` in this case since it
attempts to find a ns that contains #{foo bar} in its set of _public_
vars.

The solution implemented here is to try the alias search twice, once
with all symbols, and again without (var symbol) forms.

This does mean that Slamhound will be unable to find aliases that are
used to refer to both public and private vars via #':

    (def foo #'b/private-var)
    (def foo #'a/public-var)    ; Just use a/public-var