Add :rename failure type to regrow
If an old ns contains a :rename entry for a missing refer, we attempt
this rename before any others.

If the old ns does not contain a possible :rename entry, the :rename
path is __NEVER__ attempted since a missing reference could be
hypothesized to be a rename of ANY public var.

slam.hound.regrow/candidate has been made private as the function
signature has changed, decreasing its suitability for public
consumption.