Keywordize ns reference keys
A common confusion regarding the ns macro is the use of symbols vs
keywords for `:require`, `:import`, etc.

The fact that both work just fine in the ns form does not help at all.

Since one goal of Slamhound is to promote a single conventional style
for ns forms, it is a good idea to be graceful on this issue.

Closes #73