Add slam.hound.test.util/with-transform-test
In order to accurately test Slamhound's interaction with files, we must
test the output of the entire process:

    - Open a reader on a file
    - Receive Clojure forms from reader (with reader macros expanded)
    - Reconstruct namespace
    - Write the new namespace back to the file, but copy everything else
      back verbatim

Many tests within Slamhound use (StringReader. (str '(literal forms)))
in order to avoid the drudgery of doing a manual pr-str.

Unfortunately, (str '[^{:my-metadata "value"} x]) outputs: "[x]", so any
metadata on the forms is irretrievably lost.

Since the interaction of Slamhound and metadata on forms is non-obvious,
I think writing our integration tests as comparisons of input and output
_files_ will help us detect these issues sooner in the future.¹

I have placed with-transform-test and with-tempfile into a common
testing namespace so that they can be used across all test files.

RE: the name `with-transform-test`:

    - I read it as "With transform, test …"
    - Many editors commonly treat ^with-.* as a specially indented form
    - It's awkward, so if a better name is found, let's use it

¹ See issue #72, and commit 54cba5c67fe0004378ea4e381ce9df800b510b63 for
  past issues with metadata