Allow asplode to accept any type implementing clojure.java.io/Coercions
The reader passed to asplode is wholly consumed by asplode and is not
used again within reconstruct. Therefore, it makes sense to contain the
entire use of the reader within asplode.

Since the most straightforward approach to creating a reader from an
input is to call clojure.java.io/reader, asplode can share the same
convenient input signature.