Bind *ns* to file's ns during read
Commit 54cba5c67fe0004378ea4e381ce9df800b510b63 introduced a method of
detecting unqualified, unquoted symbols within syntax quotes that are
consumed in-namespace.

From the commit message:

> We can avoid this situation by walking the body and de-qualifying only
> the symbols that have been qualified in the current ns. This should
> not be problematic since unknown symbols are qualified in the current
> ns by default.

Now, the "current ns" is determined by the reader to be the value of
the *ns* var. When using Slamhound from an editor plugin, it can be
assumed that *ns* is set to the namespace of the current open buffer.
Unfortunately, when using Slamhound from a command line, *ns* is
actually set to the `user` ns.

This causes symbols that we expect to be qualified as `my.ns/foo` to
actually be qualified as `user/foo`. There have been scattered reports
of slamhound attempting to alias namespaces as `user`¹. In at least some
of these cases, this problem is likely to blame.

¹ #65 comes to mind